[TensorPipe] Ignore expected errors
Pull Request resolved: https://github.com/pytorch/pytorch/pull/39182

When the TensorPipe context is closed and joined, all pending callbacks are invoked with an error of type PipeClosedError. This is normal and expected, and should not be logged.

There is still one log that should be addressed, namely when an incoming pipe from a remote worker dies after we have joined, which I still need to address. That will require some type of "signal" from the remote worker that the shutdown is intentional, for example sending an empty packet?
ghstack-source-id: 104837024

Differential Revision: [D21703036](https://our.internmc.facebook.com/intern/diff/D21703036/)