[TensorPipe] Use PrefixStore to avoid conflicting keys
Pull Request resolved: https://github.com/pytorch/pytorch/pull/39185

The TP agent used the store for two things: mapping ranks to names, and mapping names to addresses. The former was prefixed, the latter wasn't. So, if a worker had a name which was `names/0` this would lead to a conflict. We should prefix both usages, and we can do so easily with the `PrefixStore`.
ghstack-source-id: 104837023

Differential Revision: [D21767862](https://our.internmc.facebook.com/intern/diff/D21767862/)